unify the tox and github action workflows
This commit ensures that tox and github action workflows are performing
the same tests with the same tools.

Background

Currently this project uses tox for orchestrating tests from the local
development environment and uses Github actions to test from push and
pull requests.  It is important that both tools implement the same
set of tests with the same tools.

This commit unifies the two test orchestrators and adds notes to each
file to remind developers to keep them in sync.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>