adds automake setting
This commit adds a new feature switch `automake` that allows control
over whether or not the api will automatically issue the `make()`
function to auto bind to the openapi spec.  This commit introduces a new
environment setting `PUREPORT_AUTOMAKE` which accepts a boolean value.

When the setting is True (default), the api will automatically load and
bind the openapi spec as a set of functions when the api module is
imported.

Whent he setting is False, the api will not automatically bind the
functions and the responsibility is on the implementor to call the
`make()` function progammatically

This commit depends on the successful merge of #3

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>