fixes an issue when generating bindings
This commit solves an exception that gets raised when attempting to run
a bound function.  The exception complains about providing mutliple
values for the keyword `session`  This fix properly constructs the
partial function so as to not trample all over the session argument.

This commit also includes an update to the configuration setting for
automatically generating bindings.  The setting `AUTOMAKE` has been
refactored to `AUTOMAKE_BINDINGS` to better reflect its use.  The env
variable is now `PUREPORT_AUTOMAKE_BINDINGS`

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>