adds more unit test cases
This commit adds additional unit test cases to the project to increase
unit test coverage.  It uncovered one potential bug in the api module
based on the added unit test cases and adds a small fix to the api
module to convert the method to lower case before trying to find it in
the paths.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>