add support for query string parameters
This commit adds support for passing query string parameters to the api
call.  Prior to this change, there was no way to pass query strings.
This introduces a new keyword argument query` to the transport
`__call__` method for handling the query strings.  The value should be a
python dict object.  The keyword has also been extended to the `Session`
instance.

This commit also includes support for passing query stirng details from
the autogenerated api functions.  For functions that support query
parameters per the spec, the function will accept a new keywork argument
`query` as a python dict and pass the query information to the request
object.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>