refactor api bindings (#24)
This commit refactors the entire api bindings functions.  It now binds
the functions and queries to a session object instead of globally in the
api module.  The models are still bound globally.   This change allows
for per-session functions instead of a single set of functions.

This change also introduces automatically discovering the account id
based on the session.  It also will now inject the account id into
functions that require it.

Note: this is a breaking change for the api bindings.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>