updates status code check for POST operation (#33)
This commit adds status code 200 to the list of status codes that get
checked when a POST operation is called.  There are some endpoints such
as stop, start, and respawn for gateways that will return a 200 status
code.  Prior to this change, the POST operation would complete
succesfully but the client would error due to the 200 code not being
considered a valid return code for success

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>