properly handle uniqueItems (#36)
This commit changes how arrays are processed to handle both cases of
uniqueItems being set to true or false.  Prior to this change, arrays
where only converted to object types when uniqueItems was set to true.
Now array items are processed and then uniqueness is validated.

This change also includes a minor change to the functions module to
evalue the regular expression as a raw string.  This change silences a
warning produced by flake8.  No code changes were introduced to
functions.py

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>