fixes the `serialize()` method (#37)
This commit fixes an issue in the `serialize()` method such that it now
will recursively serialize the entire object.  Prior to this change, the
serialize function did not consider list items to be objects so it did
not attempt to serialize them.  The method now recursively serializes
items that are objects.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>