fixes issues managing connections (#41)
This commit address two problems and provides fixes for them.

The first issue is handling models that have discriminators.  The code
path expected the model to be in args not kwargs.  This fix now finds
the model in kwargs and handles the discriminator properly.

The second issue is with the expected status code returned during
deletes.  Deletes where not checking for a return of 200 only 204.  This
is now corrected and deletes pass on 200 or 204.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>