fix up how requests are handled (#43)
This commit fixes an issue with how requests are handled when calling
the appropriate function from ullib3.  With this change, there is more
control given to the session instance to determine how fields are
handled between body and query string.  This commit also adresses a
probem that would surface if both body and query string were passed as
kwargs.

This commit also removes a lingering debug command from the functions
code.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>